+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Update allowed classification fields to 2020-01-28 version (#1409, #1476).
   - The type `compromised` has been renamed to `system-compromise`.
   - The type `unauthorized-command` has been merged into `system-compromise`.
   - The type `unauthorized-login` has been merged into `system-compromise`.
+  - The type `defacement` has been merged into taxonomy `information-content-security`, type `unauthorised-information-modification` (PR#1994 by Sebastian Wagner, addresses #1409).
 - The taxonomy `information gathering` has been rename to `information-gathering`.
 - The taxonomy `malicious code` has been renamed to `malicious-code`.
   - The type `c2server` has been renamed to `c2-server`.